services/mcp-server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.13.3
redis[hiredis]==5.1.1
python-jose[cryptography]==3.3.0
structlog==24.4.0
prometheus-fastapi-instrumentator==7.0.0
httpx==0.27.2
pgvector>=0.3.6
cryptography==43.0.3
python-multipart==0.0.12
twilio==9.3.7
sentry-sdk[fastapi]==2.19.2
opentelemetry-sdk>=1.40.0
opentelemetry-exporter-otlp>=1.40.0
opentelemetry-instrumentation-fastapi>=0.49b0
opentelemetry-instrumentation-sqlalchemy>=0.49b0
opentelemetry-instrumentation-redis>=0.49b0
passlib[bcrypt]==1.7.4
jsonschema>=4.20.0
psycopg2-binary==2.9.9
stripe>=7.0.0
python-dotenv>=1.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7,<4.0
